AI-desktop assistant/requirements.txt
speechrecognition
wikipedia
openai
pyaudio
pyttsx3
pocketsphinx
google-generativeai